requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP & scraping
python-docx==1.1.0
rapidfuzz==3.5.2
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
lxml==4.9.3

# Data & documents
PyYAML==6.0.1
reportlab==4.0.7
PyPDF2==3.0.1

# LLM APIs
openai>=1.3.0
google-generativeai>=0.3.1

# Optional & Helper dependencies
langchain-openai
langchain-google-genai
langchain-core
langchain-community
langchain-text-splitters
python-dotenv
loguru

# Heavy dependencies (required for full matching logic)
selenium==4.15.2
webdriver-manager==4.0.1
# playwright is difficult on Heroku without custom buildpacks, keeping commented
# faiss-cpu>=1.9.0.post1 (may cause build issues on Heroku, keeping commented for now)
